+++ source/extensions/omni.flux.property_widget_builder.model.usd/omni/flux/property_widget_builder/model/usd/tests/__init__.py
@@ -35,6 +35,7 @@ from .unit.test_multichannel_paste import TestMultichannelPaste
 from .unit.test_relationship_item import TestUSDRelationshipItem
 from .unit.test_relationship_utils import TestRelationshipUtils
 from .unit.test_relationship_value_model import TestUsdRelationshipValueModel
+from .unit.test_setup_ui import TestUSDPropertyWidget
 from .unit.test_usd_attribute_item_bounds import TestUSDAttributeItemBounds
 from .unit.test_value_tooltips import TestUSDAttributeValueTooltips
 
@@ -69,6 +70,7 @@ __all__ = [
     "TestUSDAttributeValueTooltips",
     "TestUSDListener",
     "TestUSDModelInteractiveNotices",
+    "TestUSDPropertyWidget",
     "TestUSDRelationshipItem",
     "TestUsdAttributeValueModelEditBatching",
     "TestUsdGradientWidgetDragLifecycle",